requirements.txt
# Web3 and Blockchain
web3==6.11.0
eth-account==0.10.0
eth-utils==2.3.1

# Data Analysis
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Price Data and APIs
requests==2.31.0
aiohttp==3.9.1
ccxt==4.1.64

# Optimization
scikit-learn==1.3.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging and Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Backtesting
matplotlib==3.8.2
seaborn==0.13.0

# Type hints
typing-extensions==4.9.0
